fix: fix `splitRect` when rect has `x1` or `y1`
